pyproject.toml
[project]
name = "posthog-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=1.51.0",
    "fastapi[standard]>=0.115.8",
    "openai>=1.63.2",
    "pillow>=11.1.0",
    "posthog>=3.15.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "svix>=1.59.2",
    "tenacity>=9.0.0",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.15.0",
    "ruff>=0.9.7",
]
